整理清除警告

+++ iOS-git-osc/Controllers/Notes/NotesView.m
@@ -263,7 +263,7 @@ static NSString * const IssueDescriptionCellId = @"IssueDescriptionCell";
     
     AFHTTPRequestOperationManager *manager = [AFHTTPRequestOperationManager GitManager];
     
-    NSString *strUrl = [NSString stringWithFormat:@"%@%@/%@/issues/%lld/notes?private_token=%@&&page=%lu", GITAPI_HTTPS_PREFIX, GITAPI_PROJECTS, _projectNameSpace, _issue.issueId, [Tools getPrivateToken], page];
+    NSString *strUrl = [NSString stringWithFormat:@"%@%@/%@/issues/%lld/notes?private_token=%@&&page=%lu", GITAPI_HTTPS_PREFIX, GITAPI_PROJECTS, _projectNameSpace, _issue.issueId, [Tools getPrivateToken], (unsigned long)page];
     
     [manager GET:strUrl
       parameters:nil
